+++ protocol/CSharp/ProtoReflection.cs
@@ -188,6 +188,22 @@ namespace Proto4z
             _idToName.Add(10014, "DictSkillEffect");
             _nameToID.Add("DictSkill", 10015);
             _idToName.Add(10015, "DictSkill");
+            _nameToID.Add("DictProp", 11000);
+            _idToName.Add(11000, "DictProp");
+            _nameToID.Add("DictModelLevel", 11001);
+            _idToName.Add(11001, "DictModelLevel");
+            _nameToID.Add("DictModel", 11002);
+            _idToName.Add(11002, "DictModel");
+            _nameToID.Add("AOESearch", 11003);
+            _idToName.Add(11003, "AOESearch");
+            _nameToID.Add("DictBuffEffect", 11004);
+            _idToName.Add(11004, "DictBuffEffect");
+            _nameToID.Add("DictBuff", 11005);
+            _idToName.Add(11005, "DictBuff");
+            _nameToID.Add("DictSkillEffect", 11006);
+            _idToName.Add(11006, "DictSkillEffect");
+            _nameToID.Add("DictSkill", 11007);
+            _idToName.Add(11007, "DictSkill");
             _nameToID.Add("SceneClientPulse", 50000);
             _idToName.Add(50000, "SceneClientPulse");
             _nameToID.Add("AttachSceneReq", 50001);
