+++ protocol/DictCommon.xml.cache
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF - 8"?>
 
-<cppmd5>d8a8cf9375c7b062ec5e294d46c354c7</cppmd5> 
-<luamd5>d17a3d53406b319230f99805aa88b2c6</luamd5> 
-<csharpmd5>a410796e093caf5aa856c7e444ae0155</csharpmd5> 
-<xmlmd5>856812cfe92d6549b28ec91e9e7afabc</xmlmd5> 
+<cppmd5>e11cc8753822c69e10934fda819c4c35</cppmd5> 
+<luamd5>d898a355dd36ea39b63ba9ff219e7e97</luamd5> 
+<csharpmd5>f6af1a3644eb7642cb4c2d93a09eda74</csharpmd5> 
+<xmlmd5>f3b915bfb56dee78465b8010579141e4</xmlmd5> 
 <cacheNumber>
     <cache key = "AOESearch" Number = "10011" /> 
     <cache key = "DictBuff" Number = "10013" />
