+++ protocol/DictScene.xml.cache
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF - 8"?>
+
+<cppmd5>99df76bfaafdac5d62374a8392cc932a</cppmd5> 
+<luamd5>dba093201d874e7999ea1f07b38d7ea8</luamd5> 
+<csharpmd5>e760c5cd7b5cd4a09c88d7d8d1908fb9</csharpmd5> 
+<xmlmd5>940e7e564b268b13df1c1fc376b2f116</xmlmd5> 
+<cacheNumber>
+    <cache key = "AOESearch" Number = "11003" /> 
+    <cache key = "DictBuff" Number = "11005" /> 
+    <cache key = "DictBuffEffect" Number = "11004" /> 
+    <cache key = "DictModel" Number = "11002" /> 
+    <cache key = "DictModelLevel" Number = "11001" /> 
+    <cache key = "DictProp" Number = "11000" /> 
+    <cache key = "DictSkill" Number = "11007" /> 
+    <cache key = "DictSkillEffect" Number = "11006" /> 
+</cacheNumber>
