+++ protocol/ProtoSceneCommon.xml
@@ -38,32 +38,13 @@
 
 
 
-
-    <struct name="EntityModel" store="MyISAM" desc="">
-    	<member name="eid" type="ui64"   desc=""/>
-    	<member name="modelID" type="ui64" tag="key"   desc=""/>
-        <member name="modelName" type="string"   desc=""/>
-        <member name="avatarID" type="ui64"   desc=""/>
-        <member name="avatarName" type="string"   desc=""/>
-        <member name="iconID"  type="ui64" desc="头像"/>
-        <member name="level"  type="double" desc="等级"/>
-        <member name="exp"  type="double" desc="经验"/>
-        <member name="gold"  type="double" desc="金币"/>
-        <member name="camp"  type="ui16" desc=""/>
-        <member name="etype"  type="ui16" desc=""/>
-        <member name="state"  type="ui16" desc=""/>
-    </struct>
-    <array name="EntityModelArray" type="EntityModel" desc="" />
-    <map name="EntityModelMap" key="ui64" value="EntityModel" desc="" />
-
-
     <struct name="SceneGroupAvatarInfo" desc="">
         <member name="areaID" type="ui64"  desc="考虑混服情况,必须指定该玩家来自哪个当前区"/>
-        <member name="model" type="EntityModel"  desc="玩家基础数据"/>
-        <member name="fixedProps" type="DictProp"   desc="基础固定属性"/>
-        <member name="growthProps" type="DictProp"   desc="成长基础属性"/>
-        <member name="growths" type="DictProp"   desc="成长系数"/>
-
+        <member name="modelID" type="ui64"  desc="模型ID"/>
+        <member name="equips" type="DictArrayKey"   desc="额外装备"/>
+        <member name="avatarID" type="ui64"   desc=""/>
+        <member name="avatarName" type="string"   desc=""/>
+        <member name="rankScore" type="double"   desc="天梯分数"/>
         <member name="powerType" type="ui16"  desc="组队权限: 0普通,1leader,2master"/>
         <member name="token" type="string"  desc="scene服务器的口令, 该字段在广播给客户端时需要清空非自己所属的token,否则将会造成token公开."/>
     </struct>
@@ -207,11 +188,8 @@
 
     <!-- 一个实体的完整同步数据 -->
     <struct name="EntityFullData" desc="EntityFullData">
-        <member name="model" type="EntityModel"  desc="玩家基础数据"/>
-        <member name="fixedProps" type="DictProp"   desc="基础固定属性"/>
-        <member name="growthProps" type="DictProp"   desc="成长基础属性"/>
-        <member name="growths" type="DictProp"   desc="成长系数"/>
-        <member name="props" type="DictProp"   desc="当前"/>
+        <member name="baseProps" type="DictProp"   desc="基础战斗属性"/>
+        <member name="props" type="DictProp"   desc="当前战斗属性"/>
         <member name="state" type="EntityState"  desc=""/>
         <member name="mv" type="EntityMove"  desc=""/>
         <member name="report" type="EntityReport"  desc=""/>
